created a new js file with stimulus and placed the code there, and deleted my application.js . I put the data-controller in the html.
