+++ app/views/pages/home.html.erb
@@ -1,17 +1,17 @@
-<div class="container">
+<div class="container" data-controller="homepage-web">
   <div id="page-one" class="homepage">
-<h1 class="headline" aria-label="illuminate">
-  <span class="letter" aria-hidden="true">I</span>
-  <span class="letter" aria-hidden="true">l</span>
-  <span class="letter" aria-hidden="true">l</span>
-  <span class="letter" aria-hidden="true">u</span>
-  <span class="letter" aria-hidden="true">m</span>
-  <span class="letter" aria-hidden="true">i</span>
-  <span class="letter" aria-hidden="true">n</span>
-  <span class="letter" aria-hidden="true">a</span>
-  <span class="letter" aria-hidden="true">t</span>
-  <span class="letter" aria-hidden="true">e</span>
-</h1>
+    <h1 class="headline" aria-label="illuminate">
+      <span class="letter" aria-hidden="true">I</span>
+      <span class="letter" aria-hidden="true">l</span>
+      <span class="letter" aria-hidden="true">l</span>
+      <span class="letter" aria-hidden="true">u</span>
+      <span class="letter" aria-hidden="true">m</span>
+      <span class="letter" aria-hidden="true">i</span>
+      <span class="letter" aria-hidden="true">n</span>
+      <span class="letter" aria-hidden="true">a</span>
+      <span class="letter" aria-hidden="true">t</span>
+      <span class="letter" aria-hidden="true">e</span>
+    </h1>
 
     <h2 class="text-center tagline">Enhancing the beauty journey for all.</h2>
     <button type="button" class="btn btn-dark rounded-sm learn-btn">Learn more</button>
